refactor: use ClampedMessageItemBody in TextMessageItemBody
